requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.1.0
cycler==0.11.0
dnspython==2.3.0
filelock==3.12.4
fonttools==4.42.0
fsspec==2023.9.2
huggingface-hub==0.17.3
idna==3.4
imageio==2.31.1
Jinja2==3.1.2
kiwisolver==1.4.4
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
packaging==23.1
pandas==2.1.1
Pillow==10.0.0
pymongo==4.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
scikit-image==0.21.0
scipy==1.11.1
six==1.16.0
soupsieve==2.4.1
sympy==1.12
tifffile==2023.7.18
tokenizers==0.14.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.65.0
transformers==4.34.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.3
